Remove mix firmware.push
`firmware.push` had a number of issues due to it using Erlang's built-in
ssh implementation and not the system's OpenSSH. The main ones were that
it didn't know about `.ssh/config`, `ssh-agent`, how to ask for a
key passphrase, or how to keep track of host keys. Work could be done to
fix all of these issues, but the easier solution for now is to shift use
over to the shell script since it just calls the system's ssh with
appropriate arguments.

Calling `mix firmware.push` prints a short error message now for
instructions on what to do.